snapcraft/wrappers/gpu-2404-custom-wrapper: adjust XDG_DATA_DIRS env variable
We have to adjust XDG_DATA_DIRS variable to make nvidia-container-toolkit
happy and find necessary configuration files. Actually, XDG_DATA_DIRS environment
variable value we provided with is already correct (it's getting prepared for us in
https://github.com/canonical/mesa-2404/blob/main/scripts/bin/gpu-2404-provider-wrapper.in
and
https://git.launchpad.net/~canonical-kernel-snaps/canonical-kernel-snaps/+git/kernel-snaps-u24.04/tree/hooks/kernel-gpu-2404-provider-mangler?h=pc-components)
but for some reason nvidia-container-toolkit library expects these path to be relative to driver root,
which leads to a situation when driver root is getting appended twice to a path and files
can not be found. One possible (and clean) solution would be to explicitly
specify config files seach path with nvcdi.WithConfigSearchPaths() but it doesn't
work because of a bug https://github.com/NVIDIA/nvidia-container-toolkit/pull/847.

So the easiest way I found is to adjust XDG_DATA_DIRS variable in that weird way
so when nvidia-container-toolkit prepend driverRoot to it a final path became
valid and everything works.

Signed-off-by: Alexander Mikhalitsyn <aleksandr.mikhalitsyn@canonical.com>
(cherry picked from commit 5e9629b157e76aedbb5145541cc2ff872bb80261)